python_flask/requirements.txt
Flask==2.2.3
mysql-connector-python==8.0.26
pymongo==3.12.0

